refactor(aws): split in multiples classes
